Fix docstrings
