+++ src/otx/cli/cli.py
@@ -17,7 +17,7 @@ from rich.console import Console
 from otx import OTX_LOGO, __version__
 from otx.cli.utils.help_formatter import CustomHelpFormatter
 from otx.cli.utils.jsonargparse import get_short_docstring, patch_update_configs
-from otx.core.utils import get_otx_root_path
+from otx.core.utils.imports import get_otx_root_path
 
 if TYPE_CHECKING:
     from jsonargparse._actions import _ActionSubCommands
