+++ src/otx/engine/engine.py
@@ -1,6 +1,6 @@
 """Module for OTX engine components."""
 
-# Copyright (C) 2023 Intel Corporation
+# Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 
 from __future__ import annotations
