+++ src/otx/engine/utils/__init__.py
@@ -1,4 +1,4 @@
 """Util API for OTX Engine."""
 
-# Copyright (C) 2023 Intel Corporation
+# Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
