+++ tests/integration/api/__init__.py
@@ -1,2 +1,2 @@
-# Copyright (C) 2023 Intel Corporation
+# Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
